backend/requirements.txt
alembic==1.13.1
amqp==5.3.1
annotated-types==0.7.0
anyio>=3.7.1,<4.0.0
attrs==25.3.0
bcrypt==5.0.0
billiard==4.2.2
celery==5.3.4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.1
distro==1.9.0
ecdsa==0.19.1
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
mcp==1.15.0
mcp-server==0.1.4
openai==1.3.7
packaging==25.0
passlib==1.7.4
prompt_toolkit==3.0.52
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.20
redis==5.0.1
referencing==0.36.2
requests==2.31.0
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
sse-starlette==3.0.2
starlette==0.27.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
vine==5.1.0
wcwidth==0.2.14

# ===== DEVELOPMENT & TESTING DEPENDENCIES =====
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# ===== DATABASE TESTING =====
pytest-postgresql>=5.0.0
